Playwright running remotely

+++ get_metadata.sh
@@ -1,8 +1,16 @@
+# First, check if git is on the machine... if not, just use
+# the metadata file passed in.
+sha=`git rev-parse HEAD`
+if [ $? -ne 0 ]; then
+    echo 'Git not found, using existing metadata file.'
+    exit 0
+fi
+
 # Deletes existing meta file
 rm -f meta.txt
 
 # Read commit hash
-git rev-parse HEAD >> meta.txt
+echo $sha >> meta.txt
 
 # Read log message
 git log --oneline -1 >> meta.txt
